fix: init with non local storage
