+++ packages/tego/src/constants.ts
@@ -1,7 +1,11 @@
 import path from 'node:path';
 
-export const DEFAULT_DEV_PLUGINS_PATH = path.resolve('plugins', 'dev');
-export const DEFAULT_REMOTE_PLUGINS_PATH = path.resolve('plugins', 'remote');
-export const DEFAULT_BUILTIN_PLUGINS_PATH = path.resolve('plugins', 'builtin');
-export const DEFAULT_BUILTIN_PLUGINS_RELATIVE_PATH = path.join('plugins', 'builtin');
+import { parseEnvironment } from './utils';
+
+// 解析环境变量
+parseEnvironment();
+
+export const DEFAULT_DEV_PLUGINS_PATH = path.join(process.env.TEGO_RUNTIME_HOME, 'plugins', 'dev');
+export const DEFAULT_REMOTE_PLUGINS_PATH = path.join(process.env.TEGO_RUNTIME_HOME, 'plugins', 'remote');
+export const DEFAULT_BUILTIN_PLUGINS_PATH = path.join(process.env.TEGO_RUNTIME_HOME, 'plugins', 'builtin');
 export const DEFAULT_WEB_PACKAGE_NAME = '@tego/web';
